pyproject.toml
[project]
name = "generatecodesforprinting"
version = "0.1.0"
description = "CLI tool for generating print-ready A4 PDFs with QR codes and barcodes"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.0",
    "pillow>=12.0.0",
    "python-barcode[images]>=0.16.1",
    "pyyaml>=6.0.3",
    "qrcode[pil]>=8.2",
    "reportlab>=4.4.4",
]

[project.scripts]
generate = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
